Fix encrypted data for crypto fixture
The ciphertext for '.items | .[1].encrypted.data' was errornously
padded, which caused Go tests to fail on encryption tests (the
decryption was unaffected as the padding was removed either way).

The ciphertext was (in bytes):

  0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15

Which got padded to:

  0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 16 16 16 16 16 16 16 16 16 16
16 16 16 16 16

And such data was encrypted in the old version. This CL changes the
encrypted data to the one created for the ciphertext without padding.