Prefix duplicate keys
We don't need the values of the comments in the duplicate keys for `post_apps` and `cipher`, and they can actually break some JSON libraries as the JSON specification doesn't really define any behavior with regards to duplicate keys. Since we don't need them to be machine-readable, just prefix with an underscore to avoid key clashes.