Improve YAML AST structure validation function, making its input requirements stricter.
Moves the null check from validateStructure function to the more narrowly scoped validateMapItems function.
Adds commentary to explain why value for a key can be null in a map.
Moves from tracking level to tracking parentKeys as we traverse so we can provide useful debug information when the structure is invalid.