Fix heartbeats schema definition
The previous schema defined a boolean type for the query.heartbeats
field. However, in actuality when these events are generated the type is
actually a string with either the value "true" or "false". This is
because the value is extracted from URL query parameters and no type
conversion is done.

This fix allows the realtime tests to pass, as the ably-common submodule
has been updated as part of https://github.com/ably/realtime/pull/4268,
for https://ably.atlassian.net/browse/REA-1348.